Select/unselect weeks in a grid
Listen to Escape key on the grid div to deselect
all selections.
Use tabIndex prop to make <div> focusable to be able to listen to
onKeyDown events on it.
Set tabIndex="0" to make element focusable with key/tab navigation or
tabIndex to a negative number to make element focusable only via mouse
click or programmatically.