Fix staggering of horizontal lane markers
Bug:
Horizontal lane markers were staggering / aligned in a step-wise
fashion due to cycling step size being bigger by 1 than resolutionX.

What we want:
Horizontal lane markers to be on the first column, always be the first
elem of each row.

To fix it, make index of grid content elements (i, in the nominator)
0-based (revert back to original 0-based indexing) and make step size to
be equal to resolutionX.

Another way of seeing it:
Offset by 1 the start of the first row then step over it with
step size equal to resolutionX. This way 1st content element of each row
(week) will be `resolutionX`-th element (and on the 1st row it will be
0th elem). When we start with 1, resolutionX-th elem ends up being the
last elem of each row.

This way the index of the second elem of the first row would be 1 and
index of 1st elem of 2nd row will be `resolutionX` (the length of the
content row, i.e. number of weeks) and so on.